PEP8
In regard to: "This style system was extracted from a much larger Python project and so the current naming conventions, coding style, and package structure is not ideal. The conventions were those of the larger project and have not been updated yet to be PEP8/MPL compliant".

If desired, will also update coding style and convention according to pylint.